Plutus.ChainIndex.Config

Documentation

data ChainIndexConfig

Constructors

ChainIndexConfig

Fields

cicSocketPath :: String

cicDbPath :: String

cicPort :: Int

cicNetworkId :: NetworkId

cicSecurityParam :: Int

The number of blocks after which a transaction cannot be rolled back anymore

cicSlotConfig :: SlotConfig

cicStoreFrom :: BlockNo

Only store transactions from this block number onward

cicAppendTransactionQueueSize :: Natural

The size of the queue and a number of transactions to collect before writing to the database

newtype DecodeConfigException

Constructors

DecodeConfigException String

defaultConfig :: ChainIndexConfig

These settings work with the main testnet

Lenses

socketPath :: Lens' ChainIndexConfig String

dbPath :: Lens' ChainIndexConfig String

port :: Lens' ChainIndexConfig Int

networkId :: Lens' ChainIndexConfig NetworkId

securityParam :: Lens' ChainIndexConfig Int

slotConfig :: Lens' ChainIndexConfig SlotConfig

storeFrom :: Lens' ChainIndexConfig BlockNo

appendTransactionQueueSize :: Lens' ChainIndexConfig Natural

Orphan instances

FromJSON NetworkId
For some reason these are not defined anywhere, and these are the reason for the -Wno-orphans option.

Instance details

Methods

parseJSON :: Value -> Parser NetworkId

parseJSONList :: Value -> Parser [NetworkId]

FromJSON BlockNo

Instance details

Methods

parseJSON :: Value -> Parser BlockNo

parseJSONList :: Value -> Parser [BlockNo]

FromJSON NetworkMagic

Instance details

Methods

parseJSON :: Value -> Parser NetworkMagic

parseJSONList :: Value -> Parser [NetworkMagic]

ToJSON NetworkId

Instance details

Methods

toJSON :: NetworkId -> Value

toEncoding :: NetworkId -> Encoding

toJSONList :: [NetworkId] -> Value

toEncodingList :: [NetworkId] -> Encoding

ToJSON BlockNo

Instance details

Methods

toJSON :: BlockNo -> Value

toEncoding :: BlockNo -> Encoding

toJSONList :: [BlockNo] -> Value

toEncodingList :: [BlockNo] -> Encoding

ToJSON NetworkMagic

Instance details

Methods

toJSON :: NetworkMagic -> Value

toEncoding :: NetworkMagic -> Encoding

toJSONList :: [NetworkMagic] -> Value

toEncodingList :: [NetworkMagic] -> Encoding