Plutus.ChainIndex.Lib

Description

Using the chain index as a library.

A minimal example that just syncs the chain index:

withDefaultRunRequirements $ runReq -> do

    syncHandler <- defaultChainSyncHandler runReq

    syncChainIndex defaultConfig runReq syncHandler

    void getLine

Documentation

data RunRequirements

The required arguments to run the chain index effects.

Constructors

RunRequirements

Fields

trace :: Trace IO (PrettyObject ChainIndexLog)

stateTVar :: TVar ChainIndexState

pool :: Pool Connection

securityParam :: Int

withRunRequirements :: Configuration -> ChainIndexConfig -> (RunRequirements -> IO ()) -> IO ()

Generate the requirements to run the chain index effects given logging configuration and chain index configuration.

withDefaultRunRequirements :: (RunRequirements -> IO ()) -> IO ()

Generate the requirements to run the chain index effects given default configurations.

defaultLoggingConfig :: IO Configuration

The default logging configuration.

defaultConfig :: ChainIndexConfig

These settings work with the main testnet

Chain index effects

handleChainIndexEffects :: forall (effs :: [Type -> Type]) a. (LastMember IO effs, Member (LogMsg ChainIndexLog) effs) => RunRequirements -> Eff (ChainIndexQueryEffect ': (ChainIndexControlEffect ': (BeamEffect Sqlite ': effs))) a -> Eff effs (Either ChainIndexError a)

Handle the chain index effects from the set of all effects.

runChainIndexEffects :: RunRequirements -> Eff '[ChainIndexQueryEffect, ChainIndexControlEffect, BeamEffect Sqlite] a -> IO (Either ChainIndexError a)

Run the chain index effects.

Chain synchronisation

syncChainIndex :: ChainIndexConfig -> RunRequirements -> ChainSyncHandler -> IO ()

Synchronise the chain index with the node using the given handler.

Synchronisation handlers

type ChainSyncHandler = ChainSyncEvent -> IO ()

A handler for chain synchronisation events.

data ChainSyncEvent

Chain synchronisation events.

Constructors

Resume Point
Resume from the given point

RollForward ChainSyncBlock Tip
Append the given block. The tip is the current tip of the node, which is newer than the tip of the block during syncing.

RollBackward Point Tip
Roll back to the given point. The tip is current tip of the node.

type EventsQueue = TBMQueue ChainSyncEvent

storeChainSyncHandler :: EventsQueue -> ChainSyncHandler

storeFromBlockNo :: BlockNumber -> ChainSyncHandler -> ChainSyncHandler

Changes the given ChainSyncHandler to only store transactions with a block number no smaller than the given one.

filterTxs

Arguments

:: (ChainIndexTx -> Bool)
Only process transactions for which this function returns True.

-> (ChainIndexTx -> Bool)
From those, only store transactions for which this function returns True.

-> ChainSyncHandler
The ChainSyncHandler on which the returned ChainSyncHandler is based.

-> ChainSyncHandler

Changes the given ChainSyncHandler to only process and store certain transactions.

runChainIndexDuringSync :: RunRequirements -> Eff '[ChainIndexQueryEffect, ChainIndexControlEffect, BeamEffect Sqlite] a -> IO (Maybe a)

Utils

getTipSlot :: ChainIndexConfig -> IO (Maybe SlotNo)

Get the slot number of the current tip of the node.