Plutus.ChainIndex.SyncStats

Documentation

data SyncStats

Constructors

SyncStats

Fields

syncStatsAppliedBlocks :: Integer

Number of applied blocks

syncStatsAppliedRollbacks :: Integer

Number of rollbacks

syncStatsChainSyncPoint :: Point

Chain index syncing tip

syncStatsNodePoint :: Point

Current tip of the node

data SyncLog

Constructors

SyncLog

Fields

syncStateSyncLog :: SyncState

State of the syncing

syncStatsSyncLog :: SyncStats

Stats of the syncing

syncPeriodSyncLog :: Second

Period in seconds used to accumulate log events

data SyncState

Constructors

Synced

Syncing Double

NotSyncing

isSyncStateSynced :: SyncState -> Bool

isSyncStateNotSyncing :: SyncState -> Bool

isSyncStateSyncing :: SyncState -> Bool

logProgress :: forall effs. Member (LogMsg SyncLog) effs => [ChainSyncEvent] -> TimeSpec -> Eff effs ()

Log syncing summary.

getSyncStateFromStats :: SyncStats -> SyncState

Get the SyncState for a SyncState.

TODO: The syncing percentage is valid when the node is already fully synced. But when the node and chain-index are started at the same time, the syncing percentage is not a valid number considering the actual tip of the node. Find a better way to calculate this percentage.

getSyncState :: Point -> Point -> SyncState

convertEventToSyncStats :: ChainSyncEvent -> SyncStats